requirements.txt
python_barcode==0.16.1
pillow==11.3.0